refactor(claude): merged redundant query generator class into ClaudeAssistant
